refactor: remove map service
